+++ CHANGELOG.md
@@ -4,6 +4,15 @@ Notable changes to Ascent are documented in this file. This changelog started on
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project aspires to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased 
+### Preferred dependency versions for ascent@develop
+- conduit@0.8.8
+- vtk-m@2.0.0
+
+### Added
+- Added parameters to control HDF5 compression options to the Relay Extract.
+
+
 ## [0.9.2] - Released 2023-06-30
 ### Preferred dependency versions for ascent@0.9.2
 - conduit@0.8.8